+++ ICD.Connect.Misc.CrestronPro/Devices/OccupancySensors/CenOdtCPoeAdapter.cs
@@ -24,7 +24,9 @@ namespace ICD.Connect.Misc.CrestronPro.Devices.OccupancySensors
 	{
 		#region Events
 
+		public event EventHandler<GenericEventArgs<eOccupancyFeatures>> OnSupportedFeaturesChanged;
 		public event EventHandler<GenericEventArgs<eOccupancyState>> OnOccupancyStateChanged;
+		public event EventHandler<IntEventArgs> OnPeopleCountChanged;
 
 		#endregion
 
@@ -40,6 +42,8 @@ namespace ICD.Connect.Misc.CrestronPro.Devices.OccupancySensors
 
 		#region Properties
 
+		public eOccupancyFeatures SupportedFeatures { get { return eOccupancyFeatures.Occupancy; } }
+
 		public eOccupancyState OccupancyState
 		{
 			get { return m_OccupancyState; }
@@ -53,6 +57,8 @@ namespace ICD.Connect.Misc.CrestronPro.Devices.OccupancySensors
 				OnOccupancyStateChanged.Raise(this, new GenericEventArgs<eOccupancyState>(m_OccupancyState));
 			}
 		}
+
+		public int PeopleCount { get { return 0; } }
 
 		#endregion