+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+ - Occupancy sensors implement new IOccupancySensorControl SupportedFeatures and PeopleCount
+
 ## [10.0.1] - 2021-08-03
 ### Changed
  - CrestronEthernetDeviceUtils - don't start threads for SSH requests